pyproject.toml
[project]
name = "ratu-onchain-monitor"
version = "0.1.0"
description = "On-chain whale and token holder analytics using Ankr API"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Adityo Nugroho", email = "adityo.nugroho.id@gmail.com" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
onchain-monitor = "onchain_monitor.main:main"

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

# =============================================================================
# Pytest Configuration
# =============================================================================
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]

# =============================================================================
# Ruff Linting Configuration (LOOSER)
# =============================================================================
[tool.ruff]
exclude = [".git", ".venv", "build", "dist", "__pycache__", "*.egg-info"]
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]
ignore = ["F401", "F841", "E501"]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
